move or_fun_call to its own module
